Annotation updates (#14)
* add explicit residue numbers for ligand chain, move plip inside from_pli, remove some dependencies on plip

* add crystal contact calculation

* feat: make system_pass_validation_criteria the new TestCriteria

* feat: remove already annotated column

* add TPSA calc

* gemmi syntax fix

* fix: fixes for gemmi and biounit

* quick biounit_selection.FindResidue workaround + lint

* fix: test new criteria fails validation for test case b/c of pocket_max_alt_count

* chore: don't redownload existing and also moad is down

* feat: consider only ligand-receptor contacts and add crystal contacts as a validation criterion

* smiles check for consistency

* lint

* more test plus lint

* more lint

* adding PRD/smiles library - not being used [WIP]

* only log action if actually downloading

* rm unused lib

* fix: gemmi mapping indices

* feat: move remove_biounit functionality to System

* minor bug

* remove no longer used configs

* feat: move to atom count/fraction for crystal contacts and add 6lu7 test

* feat: drop ic50 from bindingdb

* feat: add has_binding_affinity property to system

* feat: add has_binding_affinity as split priority

* feat: add min max pocket and pli to split test/val

* fix: set union

* fix: split priority bug

* chore: use receptor cif instead of system cif

* feat: remove system.pdb saving

* feat: add pli_unique_qcov scoring and remove score_ligand_level functionality

* feat: add number of crystal contacted residues

* chore: lint

* add num_resolved_heavy_atoms

* save_folder Path typing

* lint

* comment out papyrus to avoid mixing IC50, EC50 data

* min_max_test_pli set to start from 3

---------

+++ src/plinder/data/pipeline/io.py
@@ -99,9 +99,9 @@ def download_cofactors(
 def download_affinity_data(
     *,
     data_dir: Path,
-    papyrus_url: str = "https://static-content.springer.com/esm/art%3A10.1186%2Fs13321-022-00672-x/MediaObjects/13321_2022_672_MOESM4_ESM.zip",
+    # papyrus_url: str = "https://static-content.springer.com/esm/art%3A10.1186%2Fs13321-022-00672-x/MediaObjects/13321_2022_672_MOESM4_ESM.zip",
     bindinddb_url: str = "https://www.bindingdb.org/bind/downloads/BindingDB_All_202401_tsv.zip",
-    moad_url: str = "http://www.bindingmoad.org/files/csv/every.csv",
+    # moad_url: str = "http://www.bindingmoad.org/files/csv/every.csv",
     force_update: bool = False,
 ) -> Any:
     """
@@ -111,12 +111,12 @@ def download_affinity_data(
     ----------
     data_dir : Path
         the root plinder dir
-    papyrus_url : str
-        papyrus url
+    # papyrus_url : str
+    #     papyrus url
     bindinddb_url : str
         bindingdb : url
-    moad_url :
-         moad url
+    # moad_url :
+    #      moad url
     force_update : bool, default=False
         if True, re-download data
 
@@ -143,25 +143,29 @@ def download_affinity_data(
     bindingdb_raw_affinity_path.parent.mkdir(parents=True, exist_ok=True)
     moad_raw_affinity_path.parent.mkdir(parents=True, exist_ok=True)
     if not affinity_path.is_file() or force_update:
-        # Download Papyrus
-        LOG.info(f"download_papyrus_affinity_data: {papyrus_url}")
-        resp = requests.get(papyrus_url)
-        resp.raise_for_status()
-        papyrus_raw_affinity_path.write_bytes(resp.content)
+        # # Download Papyrus
+        # if not papyrus_raw_affinity_path.exists():
+        #     LOG.info(f"download_papyrus_affinity_data: {papyrus_url}")
+        #     resp = requests.get(papyrus_url)
+        #     resp.raise_for_status()
+        #     papyrus_raw_affinity_path.write_bytes(resp.content)
 
         # Download BindingDB
-        LOG.info(f"download_bindingdb_affinity_data: {bindinddb_url}")
-        with urlopen(bindinddb_url) as zipresp:
-            with ZipFile(BytesIO(zipresp.read())) as zfile:
-                zfile.extractall()
-        with open(bindingdb_raw_affinity_path, "w") as out:
-            with open(Path.cwd() / "BindingDB_All_202401_tsv") as f:
-                out.write(f.read())
-        # Download MOAD
-        LOG.info(f"download_moad_affinity_data: {moad_url}")
-        r = requests.get(moad_url, allow_redirects=True)
-        with open(moad_raw_affinity_path, "wb") as out:
-            out.write(r.content)
+        if not bindingdb_raw_affinity_path.exists():
+            LOG.info(f"download_bindingdb_affinity_data: {bindinddb_url}")
+            with urlopen(bindinddb_url) as zipresp:
+                with ZipFile(BytesIO(zipresp.read())) as zfile:
+                    zfile.extractall()
+            with open(bindingdb_raw_affinity_path, "w") as out:
+                with open(Path.cwd() / "BindingDB_All_202401.tsv") as f:
+                    out.write(f.read())
+        # # Download MOAD
+        # TODO: moad URL not working
+        # if not moad_raw_affinity_path.exists():
+        #     LOG.info(f"download_moad_affinity_data: {moad_url}")
+        #     r = requests.get(moad_url, allow_redirects=True)
+        #     with open(moad_raw_affinity_path, "wb") as out:
+        #         out.write(r.content)
 
         LOG.info("transforming_affinity_data: extracting median affinity")
         binding_db_affinity_df = transform.transform_bindingdb_affinity_data(
@@ -169,19 +173,21 @@ def download_affinity_data(
         )
         binding_db_affinity_df["preference"] = 1
 
-        moad_affinity_df = transform.transform_moad_affinity_data(
-            raw_affinity_path=moad_raw_affinity_path
-        )
-        moad_affinity_df["preference"] = 2
+        # moad_affinity_df = transform.transform_moad_affinity_data(
+        #     raw_affinity_path=moad_raw_affinity_path
+        # )
+        # moad_affinity_df["preference"] = 2
 
-        papyrus_affinity_df = transform.transform_papyrus_affinity_data(
-            raw_affinity_path=papyrus_raw_affinity_path
-        )
-        papyrus_affinity_df["preference"] = 3
-        all_affinity_df = pd.concat(
-            [papyrus_affinity_df, binding_db_affinity_df, moad_affinity_df],
-            ignore_index=True,
-        ).drop_duplicates()
+        # papyrus_affinity_df = transform.transform_papyrus_affinity_data(
+        #     raw_affinity_path=papyrus_raw_affinity_path
+        # )
+        # papyrus_affinity_df["preference"] = 3
+        # all_affinity_df = pd.concat(
+        #     # [papyrus_affinity_df, binding_db_affinity_df, moad_affinity_df],
+        #     [papyrus_affinity_df, binding_db_affinity_df],
+        #     ignore_index=True,
+        # ).drop_duplicates()
+        all_affinity_df = binding_db_affinity_df.drop_duplicates()
         all_affinity_df = all_affinity_df[all_affinity_df["pchembl"].notna()]
 
         all_affinity_df = all_affinity_df.loc[
@@ -385,9 +391,10 @@ def download_kinase_data(
     kinase_uniprotac_path : Path
         location of downloaded kinase data
     """
-    LOG.info(f"download_kinase_data: data_dir={data_dir}")
     kinase_info_path = data_dir / "dbs" / "kinase" / "kinase_information.parquet"
-    kinase_info_path.parent.mkdir(exist_ok=True, parents=True)
+    if not kinase_info_path.parent.exists() or force_update:
+        LOG.info(f"download_kinase_data: data_dir={data_dir}")
+        kinase_info_path.parent.mkdir(exist_ok=True, parents=True)
     if not kinase_info_path.is_file() or force_update:
         LOG.info(f"download_kinase_data: {url}/kinase_information")
         resp = requests.get(f"{url}/kinase_information")
